restructure project folders
