+++ jest.config.ts
@@ -1,9 +1,9 @@
-const nextJest = require('next/jest');
+const nextJest = require('next/jest')
 
-const createJestConfig = nextJest({ dir: './' });
+const createJestConfig = nextJest({ dir: './' })
 
 const customJestConfig = {
   testEnvironment: 'jsdom', // Asegura que Jest use el entorno de navegador
 }
 
-module.exports = createJestConfig(customJestConfig);
+module.exports = createJestConfig(customJestConfig)
