+++ src/services/config.ts
@@ -1,0 +1,2 @@
+export const API_KEY = 'c6aeee577586ba38e487b74dfede5deb'
+export const BASE_URL = 'https://api.themoviedb.org/3'
