+++ src/styles/base/variables.scss
