+++ src/styles/components/error.module.scss
@@ -1,4 +1,4 @@
-@import '../variables'; 
+@import '../../styles/base/variables.scss'; 
 
 .errorContainer {
     margin-top: 80px;
