+++ src/styles/components/footer.module.scss
@@ -1,4 +1,4 @@
-@import '../variables';  
+@import '../../styles/base/variables.scss';  
 
 .footer {
   .footerContent {
