+++ src/styles/components/header.module.scss
@@ -1,4 +1,4 @@
-@import '../variables'; 
+@import '../../styles/base/variables.scss'; 
 
 .header {
     position: fixed;
