+++ src/styles/components/heroSection.module.scss
@@ -1,4 +1,4 @@
-@import '../variables'; 
+@import '../../styles/base/variables.scss'; 
 
 .hero {
   position: relative;
