+++ src/styles/components/home.module.scss
@@ -1,4 +1,4 @@
-@import '../variables'; 
+@import '../../styles/base/variables.scss'; 
 
 .title {
     color: $text-color;
