+++ src/styles/components/seriesCard.module.scss
@@ -1,4 +1,4 @@
-@import '../variables'; 
+@import '../../styles/base/variables.scss'; 
 
 .seriesCard {
   display: flex;
