+++ src/styles/components/seriesDetail.module.scss
@@ -1,4 +1,4 @@
-@import '../variables'; 
+@import '../../styles/base/variables.scss'; 
 
 .detailContainer {
     margin-top: 100px;
