+++ src/styles/components/seriesPagination.module.scss
@@ -1,4 +1,4 @@
-@import '../variables'; 
+@import '../../styles/base/variables.scss'; 
 
 .pagination {
     display: flex;
