+++ src/tests/components/seriesCard.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, screen, fireEvent } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import { SeriesCard } from '../components/SeriesCard'
+import { SeriesCard } from '../../components/SeriesCard'
 
 const mockSeries = {
   id: 202250,
