+++ src/tests/mocks/__mocks__/next/router.js
