+++ src/tests/pages/page.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import Home from '../../src/app/page'
+import Home from '../../../src/app/page'
 
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn().mockReturnValue({
