+++ tsconfig.json
@@ -28,7 +28,7 @@
     "src/**/*.ts",
     "src/**/*.tsx",
     ".next/types/**/*.ts"
-  ],
+, "jest.config.ts", "next.config.ts"  ],
   "exclude": [
     "node_modules"
   ]
